Fix out-of-bound accesses when getting containing record
Dereferencing on the list element means that the resultant pointer is only
valid for the bytes of the list element instead of the whole containing record,
which means that `element_ptr` yields an out-of-bounds pointer.

By deferring the dereference to after `element_ptr`, we get a valid pointer to
the containing record.